+++ packages/typespec-client-generator-core/src/types.ts
@@ -136,7 +136,15 @@ function getAnyType(context: TCGCContext, type: Type): [SdkBuiltInType, readonly
 
 function getEncodeHelper(context: TCGCContext, type: Type, kind: string): string {
   if (type.kind === "ModelProperty" || type.kind === "Scalar") {
-    return getEncode(context.program, type)?.encoding || kind;
+    const encode = getEncode(context.program, type);
+    if (encode?.encoding) {
+      return encode.encoding;
+    }
+    if (encode?.type) {
+      // if we specify the encoding type in the decorator, we set the `.encode` string
+      // to the kind of the encoding type
+      return getSdkBuiltInType(context, encode.type).kind;
+    }
   }
   return kind;
 }
@@ -249,9 +257,10 @@ function getSdkBuiltInTypeWithDiagnostics(
     encode: getEncodeHelper(context, type, kind),
     description: docWrapper.description,
     details: docWrapper.details,
-    baseType: type.baseScalar
-      ? diagnostics.pipe(getSdkBuiltInTypeWithDiagnostics(context, type.baseScalar, kind))
-      : undefined,
+    baseType:
+      type.baseScalar && !context.program.checker.isStdType(type) // we only calculate the base type when this type has a base type and this type is not a std type because for std types there is no point of calculating its base type.
+        ? diagnostics.pipe(getSdkBuiltInTypeWithDiagnostics(context, type.baseScalar, kind))
+        : undefined,
     crossLanguageDefinitionId: getCrossLanguageDefinitionId(context, type),
   };
   addEncodeInfo(context, type, stdType);
@@ -301,9 +310,10 @@ function getSdkDateTimeType(
 ): [SdkDateTimeType, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   const docWrapper = getDocHelper(context, type);
-  const baseType = type.baseScalar
-    ? diagnostics.pipe(getSdkDateTimeType(context, type.baseScalar, kind))
-    : undefined;
+  const baseType =
+    type.baseScalar && !context.program.checker.isStdType(type) // we only calculate the base type when this type has a base type and this type is not a std type because for std types there is no point of calculating its base type.
+      ? diagnostics.pipe(getSdkDateTimeType(context, type.baseScalar, kind))
+      : undefined;
   const [encode, wireType] = getEncodeInfoForDateTimeOrDuration(
     context,
     getEncode(context.program, type),
@@ -399,9 +409,10 @@ function getSdkDurationTypeWithDiagnostics(
 ): [SdkDurationType, readonly Diagnostic[]] {
   const diagnostics = createDiagnosticCollector();
   const docWrapper = getDocHelper(context, type);
-  const baseType = type.baseScalar
-    ? diagnostics.pipe(getSdkDurationTypeWithDiagnostics(context, type.baseScalar, kind))
-    : undefined;
+  const baseType =
+    type.baseScalar && !context.program.checker.isStdType(type) // we only calculate the base type when this type has a base type and this type is not a std type because for std types there is no point of calculating its base type.
+      ? diagnostics.pipe(getSdkDurationTypeWithDiagnostics(context, type.baseScalar, kind))
+      : undefined;
   const [encode, wireType] = getEncodeInfoForDateTimeOrDuration(
     context,
     getEncode(context.program, type),
@@ -1301,10 +1312,6 @@ function updateMultiPartInfo(
   if (base.multipartOptions !== undefined) {
     base.isMultipartFileInput = base.multipartOptions.isFilePart;
   }
-  if (base.multipartOptions?.isMulti && base.type.kind === "array") {
-    // for "images: T[]" or "images: HttpPart<T>[]", return type shall be "T" instead of "T[]"
-    base.type = base.type.valueType;
-  }
 
   return diagnostics.wrap(undefined);
 }
@@ -1329,7 +1336,15 @@ export function getSdkModelPropertyType(
     flatten: shouldFlattenProperty(context, type),
   };
   if (operation) {
-    diagnostics.pipe(updateMultiPartInfo(context, type, result, operation));
+    const httpOperation = getHttpOperationWithCache(context, operation);
+    if (
+      type.model &&
+      httpOperation.parameters.body &&
+      httpOperation.parameters.body.type === type.model
+    ) {
+      // only add multipartOptions for property of multipart body
+      diagnostics.pipe(updateMultiPartInfo(context, type, result, operation));
+    }
   }
   return diagnostics.wrap(result);
 }