+++ packages/typespec-client-generator-core/test/package.test.ts
@@ -313,7 +313,7 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(templateArg.optional, false);
       strictEqual(templateArg.onClient, true);
       strictEqual(templateArg.clientDefaultValue, undefined);
-      strictEqual(templateArg.description, "Testserver endpoint");
+      strictEqual(templateArg.description, undefined);
 
       const credentialParam = client.initialization.properties.filter(
         (p): p is SdkCredentialParameter => p.kind === "credential"
@@ -405,6 +405,38 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(scheme.name, "x-ms-api-key");
     });
 
+    it("endpoint with path param default value", async () => {
+      await runner.compile(`
+        @server(
+          "{endpoint}",
+          "Test server endpoint",
+          {
+            endpoint: string = "http://localhost:3000",
+          }
+        )
+        @service({})
+        namespace MyService;
+      `);
+      const sdkPackage = runner.context.sdkPackage;
+      strictEqual(sdkPackage.clients.length, 1);
+      const client = sdkPackage.clients[0];
+      strictEqual(client.initialization.properties.length, 1);
+
+      const endpointParam = client.initialization.properties.filter(
+        (p): p is SdkEndpointParameter => p.kind === "endpoint"
+      )[0];
+      strictEqual(endpointParam.type.kind, "endpoint");
+      strictEqual(endpointParam.type.serverUrl, "{endpoint}");
+
+      strictEqual(endpointParam.type.templateArguments.length, 1);
+      const endpointTemplateArg = endpointParam.type.templateArguments[0];
+      strictEqual(endpointTemplateArg.name, "endpoint");
+      strictEqual(endpointTemplateArg.onClient, true);
+      strictEqual(endpointTemplateArg.optional, false);
+      strictEqual(endpointTemplateArg.kind, "path");
+      strictEqual(endpointTemplateArg.clientDefaultValue, "http://localhost:3000");
+    });
+
     it("single with core", async () => {
       const runnerWithCore = await createSdkTestRunner({
         librariesToAdd: [AzureCoreTestLibrary],
@@ -447,6 +479,7 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(sdkPackage.clients.length, 1);
       const client = sdkPackage.clients[0];
       strictEqual(client.name, "ServiceClient");
+      strictEqual(client.crossLanguageDefinitionId, "My.Service");
       strictEqual(client.initialization.properties.length, 3);
       strictEqual(client.apiVersions.length, 1);
       strictEqual(client.apiVersions[0], "2022-12-01-preview");
@@ -516,6 +549,7 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(sdkPackage.clients.length, 1);
       const client = sdkPackage.clients[0];
       strictEqual(client.name, "ServiceClient");
+      strictEqual(client.crossLanguageDefinitionId, "My.Service");
       strictEqual(client.initialization.properties.length, 3);
       strictEqual(client.apiVersions.length, 2);
       deepStrictEqual(client.apiVersions, ["2022-12-01-preview", "2022-12-01"]);
@@ -579,6 +613,7 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(mainClient.methods.length, 1);
       strictEqual(mainClient.initialization.properties.length, 1);
       strictEqual(mainClient.initialization.properties[0].name, "endpoint");
+      strictEqual(mainClient.crossLanguageDefinitionId, "TestService");
 
       const clientAccessor = mainClient.methods[0];
       strictEqual(clientAccessor.kind, "clientaccessor");
@@ -596,6 +631,7 @@ describe("typespec-client-generator-core: package", () => {
         operationGroup.methods[0].crossLanguageDefintionId,
         "TestService.MyOperationGroup.func"
       );
+      strictEqual(operationGroup.crossLanguageDefinitionId, "TestService.MyOperationGroup");
     });
 
     it("operationGroup2", async () => {
@@ -629,6 +665,7 @@ describe("typespec-client-generator-core: package", () => {
       ok(mainClient.initialization);
       strictEqual(mainClient.initialization.properties.length, 1);
       strictEqual(mainClient.initialization.properties[0].name, "endpoint");
+      strictEqual(mainClient.crossLanguageDefinitionId, "TestService");
 
       const fooAccessor = mainClient.methods[0];
       strictEqual(fooAccessor.kind, "clientaccessor");
@@ -649,6 +686,7 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(fooClient.initialization.properties.length, 1);
       strictEqual(fooClient.initialization.access, "internal");
       strictEqual(fooClient.methods.length, 1);
+      strictEqual(fooClient.crossLanguageDefinitionId, "TestService.Foo");
 
       const fooBarAccessor = fooClient.methods[0];
       strictEqual(fooBarAccessor.kind, "clientaccessor");
@@ -660,6 +698,7 @@ describe("typespec-client-generator-core: package", () => {
 
       strictEqual(fooBarClient.initialization.properties.length, 1);
       strictEqual(fooBarClient.initialization.access, "internal");
+      strictEqual(fooBarClient.crossLanguageDefinitionId, "TestService.Foo.Bar");
       strictEqual(fooBarClient.methods.length, 1);
       strictEqual(fooBarClient.methods[0].kind, "basic");
       strictEqual(fooBarClient.methods[0].name, "one");
@@ -667,6 +706,7 @@ describe("typespec-client-generator-core: package", () => {
 
       strictEqual(barClient.initialization.properties.length, 1);
       strictEqual(barClient.initialization.access, "internal");
+      strictEqual(barClient.crossLanguageDefinitionId, "TestService.Bar");
       strictEqual(barClient.methods.length, 1);
       strictEqual(barClient.methods[0].kind, "basic");
       strictEqual(barClient.methods[0].name, "two");
@@ -3071,6 +3111,14 @@ describe("typespec-client-generator-core: package", () => {
       strictEqual(op.bodyParam.kind, "body");
       strictEqual(op.bodyParam.name, "testRequest");
       deepStrictEqual(op.bodyParam.correspondingMethodParams, [documentMethodParam]);
+
+      const anonymousModel = runner.context.sdkPackage.models[0];
+      strictEqual(anonymousModel.properties.length, 1);
+      strictEqual(anonymousModel.properties[0].kind, "property");
+      strictEqual(anonymousModel.properties[0].isMultipartFileInput, true);
+      ok(anonymousModel.properties[0].multipartOptions);
+      strictEqual(anonymousModel.properties[0].multipartOptions.isFilePart, true);
+      strictEqual(anonymousModel.properties[0].multipartOptions.isMulti, false);
     });
 
     it("anonymous model with @body should not be spread", async () => {